refactor BulkMintingCommand
