+++ packages/cli/src/app/Commands/IPFSCommand.ts
@@ -1,9 +1,14 @@
 import { NftPlaybookException, PinataClient } from '@nft-playbook/middleware';
 import { CliStrings } from '../CliStrings';
-import { Command, getInput, showException, sleep } from './Command';
+import {
+  checkPinataCredentials,
+  Command,
+  getInput,
+  showException,
+  sleep,
+} from './Command';
 import { middleware } from '@nft-playbook/middleware';
 import fs = require('fs');
-import inquirer = require('inquirer');
 import { Worker } from 'worker_threads';
 
 export class IPFSCommand implements Command {
@@ -19,45 +24,8 @@ export class IPFSCommand implements Command {
 
   async execute() {
     this.print_header();
-    if (
-      process.env.PINATA_API_KEY === undefined ||
-      process.env.PINATA_API_KEY.length === 0 ||
-      process.env.PINATA_API_SEC === undefined ||
-      process.env.PINATA_API_SEC.length === 0
-    ) {
-      if (
-        PinataClient.apiKey === null ||
-        PinataClient.apiKey === undefined ||
-        PinataClient.apiSec === null ||
-        PinataClient.apiSec === undefined
-      ) {
-        console.log(CliStrings.IPFSWarningMessage);
-        const promptQuestion: inquirer.QuestionCollection = [
-          {
-            type: 'confirm',
-            name: 'confirmed',
-            message: CliStrings.IPFSConfirmationQuestion,
-          },
-        ];
-        const answer = await inquirer.prompt(promptQuestion);
-        if (answer.confirmed) {
-          const key = await getInput(CliStrings.IPFSQuestionApiKey, '');
-          if (key === null) return;
-          PinataClient.apiKey = key;
-
-          const sec = await getInput(CliStrings.IPFSQuestionApiSec, '');
-          if (sec === null) return;
-          PinataClient.apiSec = sec;
-        } else {
-          return;
-        }
-      }
-    } else {
-      console.log(CliStrings.IPFSEnvFile);
-      PinataClient.apiKey = process.env.PINATA_API_KEY;
-      PinataClient.apiSec = process.env.PINATA_API_SEC;
-      PinataClient.apiKey = process.env.PINATA_API_KEY;
-      PinataClient.apiSec = process.env.PINATA_API_SEC;
+    if (!(await checkPinataCredentials())) {
+      return;
     }
     const path = await getInput(CliStrings.IPFSFileConfirmationQuestion, '');
     if (path === null) return;
